+++ main.js
@@ -48,3 +48,7 @@ app.on("before-quit", () => {
   app.removeAllListeners("close")
 })
 app.commandLine.appendSwitch("disable-smooth-scrolling", "true")
+
+
+
+